refactor: make daily chests follow the resource model
